feat: complete model loading

+++ _29_LoadingModels/Vulkan/Vertex/VertexDataMgr.cpp
@@ -20,7 +20,7 @@ std::vector<Vertex> VertexDataMgr::vertices =
         {{-0.5f, 0.5f, -0.5f}, {1.0f, 1.0f, 1.0f}, {1.0f, 1.0f}}
     };
 
-std::vector<uint16_t> VertexDataMgr::indices =
+std::vector<uint32_t> VertexDataMgr::indices =
     {
         0, 1, 2, 2, 3, 0,
         4, 5, 6, 6, 7, 4
